Move TECE ScaledSigmoid/ScaledSiLU into shared activation module.
Keep SO(2) gate activations architecture-agnostic alongside SmoothLeakyReLU.
